Remove dev search interface
- no longer needed with overview hooked-up

Replace old search entrypoint with new one

- refactoring commit
- start of the cleanup for integrating new search module
