+++ src/options/containers/settings/index.jsx
@@ -1,7 +1,6 @@
 import React from 'react'
 
 import Blacklist from '../../blacklist'
-import Search from '../../search'
 import Preferences from '../../preferences'
 import { routeTitle, sectionTitle } from '../../base.css'
 import styles from './style.css'
@@ -19,11 +18,6 @@ const SettingsContainer = () => (
             <h2 className={sectionTitle}>Miscellaneous Preferences</h2>
 
             <Preferences />
-        </section>
-        <section className={styles.section}>
-            <h2 className={sectionTitle}>search-index dev</h2>
-
-            <Search />
         </section>
     </div>
 )
